requirements.txt
dbt-snowflake==1.9.4
dbt-core==1.9.8
pandas==2.3.1
load-dotenv==0.1.0